+++ src/components/modal/ModalFrames/ModalBox.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 
 import styled from "styled-components";
-import { FlexAttribute } from "../../style/Mixin";
+import { FlexAttribute } from "../../../style/Mixin";
 
 import ModalWrapper from "./ModalWrapper";
-import theme from "../../style/Theme";
+import theme from "../../../style/Theme";
 
 function ModalBox({ children }) {
   useEffect(() => {
